fix: common 모듈 패키지명 변경, MongoDBConverter 클래스 수정
