+++ common/logging/src/main/kotlin/com/gc/common/logging/ServerLoggingInterceptor.kt
@@ -1,7 +1,6 @@
 package com.gc.common.logging
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import com.gc.common.utils.logger
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import org.springframework.stereotype.Component
