+++ customer/src/main/kotlin/com/gc/api/customer/framework/config/feign/FeignClientConfig.kt
@@ -11,11 +11,6 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 class FeignClientConfig {
 
-//  @Bean
-//  fun feignLoggerLevel(): Logger.Level {
-//    return Logger.Level.FULL
-//  }
-
   @Bean
   fun errorDecoder(): ErrorDecoder {
     return FeignErrorDecoder()
